requirements.txt
elevenlabs==2.10.0
feedparser==6.0.11
ics==0.7.2
ollama==0.5.3
phue==1.1
Pillow==11.3.0
pyserial==3.5
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
python_vlc==3.0.21203
pytz==2024.1
Requests==2.32.5
streamdeck==0.9.7
